backend/requirements.txt
pandas
azure-storage-blob
openpyxl
djangorestframework
djangorestframework-simplejwt
reportlab
